Some minor refactorings

+++ elfi/store.py
@@ -240,14 +240,19 @@ class OutputPool:
 
         os.makedirs(self.path, exist_ok=True)
 
+        # Change the working directory so that relative paths to the pool data folder can
+        # be reliably used. This allows moving and renaming of the folder.
+        cwd = os.getcwd()
+        os.chdir(self.path)
         # Pickle the stores separately
         for node, store in self.stores.items():
-            filename = os.path.join(self.path, node + '.pkl')
+            filename = node + '.pkl'
             try:
                 pickle.dump(store, open(filename, 'wb'))
             except:
                 raise IOError('Failed to pickle the store for node {}, please check that '
                               'it is pickleable or remove it before saving.'.format(node))
+        os.chdir(cwd)
 
         # Save the pool itself with stores replaced with Nones
         stores = self.stores
@@ -354,8 +359,9 @@ class ArrayPool(OutputPool):
 
     Notes
     -----
-    The default store is elfi.store.NpyStore that uses NpyArrays as stores. The latter is
-    a wrapper over NumPy .npy binary file for array data.
+    The default store is implemented in elfi.store.NpyStore that uses NpyArrays as stores.
+    The NpyArray is a wrapper over NumPy .npy binary file for array data and supports
+    appending the .npy file. It uses the .npy format 2.0 files.
 
     """
 
@@ -368,35 +374,6 @@ class ArrayPool(OutputPool):
 
         filename = os.path.join(self.path, node)
         return NpyStore(filename, self.batch_size)
-
-    def load_npy_file(self, node, n_batches=None):
-        """
-
-        Parameters
-        ----------
-        node : str
-            The node.npy file must be in self.path
-        n_batches : int, optional
-            How many batches to load. Default is as many as the array has length.
-
-        """
-
-        filename = os.path.join(self.path, node + '.npy')
-        if not os.path.exists(filename):
-            raise FileNotFoundError("Could not find file {}".format(filename))
-
-        self.add_store(node)
-        store = self.get_store(node)
-
-        if n_batches is None:
-            if len(store.array) % self.batch_size != 0:
-                self.remove_store(node)
-                raise ValueError('The array length is not divisible by the batch size')
-            n_batches = len(store.array) // self.batch_size
-        else:
-            store.array.truncate(n_batches*self.batch_size)
-
-        store.n_batches = n_batches
 
 
 class StoreBase:
@@ -452,8 +429,16 @@ class ArrayStore(StoreBase):
 
     This class is intended to make it easy to use objects that support array indexing
     as outputs stores for nodes.
+
+    Attributes
+    ----------
+    array : array-like
+        The array that the batches are stored to
+    batch_size : int
+    n_batches : int
+        How many batches are available from the underlying array.
     """
-    def __init__(self, array, batch_size, n_batches=0):
+    def __init__(self, array, batch_size, n_batches=-1):
         """
 
         Parameters
@@ -462,10 +447,16 @@ class ArrayStore(StoreBase):
             Any array like object supporting Python list indexing
         batch_size : int
             Size of a batch of data
-        n_batches : int
-            When using pre allocated arrays, this keeps track of the number of batches
-            currently stored to the array.
+        n_batches : int, optional
+            How many batches should be made available from the array. Default is -1
+            meaning all available batches.
         """
+
+        if n_batches == -1:
+            if len(array) % batch_size != 0:
+                logger.warning("The array length is not divisible by the batch size.")
+            n_batches = len(array) // batch_size
+
         self.array = array
         self.batch_size = batch_size
         self.n_batches = n_batches
@@ -533,16 +524,19 @@ class NpyStore(ArrayStore):
     Uses the NpyArray objects as an array store.
     """
 
-    def __init__(self, filename, batch_size, n_batches=0):
+    def __init__(self, file, batch_size, n_batches=-1):
         """
 
         Parameters
         ----------
-        filename : str
-            Path to the file, relative or absolute
+        file : NpyArray or str
+            NpyArray object or path to the .npy file
         batch_size
+        n_batches : int, optional
+            How many batches to make available from the file. Default -1 indicates that
+            all available batches.
         """
-        array = NpyArray(filename)
+        array = file if isinstance(file, NpyArray) else NpyArray(file)
         super(NpyStore, self).__init__(array, batch_size, n_batches)
 
     def __setitem__(self, batch_index, data):
@@ -654,7 +648,7 @@ class NpyArray:
             raise ValueError('Array is not opened.')
 
         if not self.initialized:
-            self._init_from_array(array)
+            self.init_from_array(array)
 
         if array.shape[1:] != self.shape[1:]:
             raise ValueError("Appended array is of different shape.")
@@ -690,7 +684,7 @@ class NpyArray:
         shape = (0,) + self.shape[1:]
         self.itemsize = np.empty(shape=shape, dtype=self.dtype).itemsize
 
-    def _init_from_array(self, array):
+    def init_from_array(self, array):
         """Initialize the object from an array.
 
         Sets the the header_length so large that it is possible to append to the array.
@@ -701,6 +695,10 @@ class NpyArray:
             Contains the oversized header bytes
 
         """
+
+        if self.initialized:
+            raise ValueError("The array has been initialized already!")
+
         self.shape = (0,) + array.shape[1:]
         self.dtype = array.dtype
         self.itemsize = array.itemsize
@@ -738,15 +736,16 @@ class NpyArray:
         -------
 
         """
-        if self.fs is None:
-            raise ValueError('Array has been deleted')
-        elif self.fs.closed:
-            raise ValueError('Array has been closed')
+        if not self.initialized:
+            raise ValueError('The array must be initialized before it can be truncated. '
+                             'Please see init_from_array.')
+
+        if self.closed:
+            raise ValueError('The array has been closed.')
 
         # Reset length
         self.shape = (length,) + self.shape[1:]
         self._prepare_header_data()
-        self._write_header_data()
 
         self.fs.seek(self.header_length + self.size*self.itemsize)
         self.fs.truncate()
